backend/requirements.txt
# python 3.9 기준
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
mysqlclient>=2.2
django-environ>=0.11

djangorestframework-simplejwt
python-dotenv
pillow

# 운영용
# gunicorn

